Extract new aws package in Elasticsearch connector
