+++ plugin/trino-elasticsearch/src/main/java/io/trino/plugin/elasticsearch/aws/AwsSecurityConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.elasticsearch;
+package io.trino.plugin.elasticsearch.aws;
 
 import io.airlift.configuration.Config;
 import io.airlift.configuration.ConfigSecuritySensitive;
