+++ plugin/trino-elasticsearch/src/test/java/io/trino/plugin/elasticsearch/aws/TestAwsSecurityConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.elasticsearch;
+package io.trino.plugin.elasticsearch.aws;
 
 import com.google.common.collect.ImmutableMap;
 import org.testng.annotations.Test;
